+++ packages/lib/src/utils/ModelPool.ts
@@ -1,11 +1,15 @@
 import type { AnyModel } from "../model/BaseModel"
 import { getModelIdPropertyName } from "../model/getModelMetadata"
-import { modelTypeKey } from "../model/metadata"
-import { isModelSnapshot } from "../model/utils"
+import { modelIdKey, modelTypeKey } from "../model/metadata"
+import { isModel, isModelSnapshot } from "../model/utils"
 import { ModelClass } from "../modelShared/BaseModelShared"
 import { getModelInfoForName } from "../modelShared/modelInfo"
 import { dataObjectParent } from "../parent/core"
-import { byModelTypeAndIdKey, getDeepObjectChildren } from "../parent/coreObjectChildren"
+import {
+  byModelTypeAndIdKey,
+  getDeepObjectChildren,
+  registerDeepObjectChildrenExtension,
+} from "../parent/coreObjectChildren"
 
 export class ModelPool {
   private pool: ReadonlyMap<string, AnyModel>
@@ -14,7 +18,7 @@ export class ModelPool {
     // make sure we don't use the sub-data $ object
     root = dataObjectParent.get(root) ?? root
 
-    this.pool = getDeepObjectChildren(root).deepByModelTypeAndId
+    this.pool = getDeepChildrenModels(getDeepObjectChildren(root))
   }
 
   findModelByTypeAndId(modelType: string, modelId: string): AnyModel | undefined {
@@ -33,3 +37,15 @@ export class ModelPool {
     return this.findModelByTypeAndId(sn[modelTypeKey], sn[modelIdPropertyName])
   }
 }
+
+const getDeepChildrenModels = registerDeepObjectChildrenExtension<Map<string, AnyModel>>({
+  initData() {
+    return new Map()
+  },
+
+  addNode(node, data) {
+    if (isModel(node)) {
+      data.set(byModelTypeAndIdKey(node[modelTypeKey], node[modelIdKey]), node)
+    }
+  },
+})